Updated NamedRoutes.java

+++ app/src/main/java/hexlet/code/util/NamedRoutes.java
@@ -9,14 +9,14 @@ public class NamedRoutes {
         return "/urls";
     }
 
-    public static String urlPath(int id) {
+    public static String urlPath(long id) {
         return urlPath(String.valueOf(id));
     }
 
     public static String urlPath(String id) {
         return "/urls/" + id;
     }
-    public static String urlChecksPath(int urlId) {
+    public static String urlChecksPath(long urlId) {
         return urlChecksPath(String.valueOf(urlId));
     }
